fix: show the chromecast devices on upgraded electron 19
