+++ www/main/viewer/utils/index.js
@@ -1,1 +1,2 @@
 export { updateViewerScale } from './ScaleViewer';
+export { castToReceiver, appendMessage, requestSession, stopApp, tingle } from './chromecast';
