+++ www/viewer.html
@@ -21,6 +21,5 @@
           root.render(app());
         });
     </script>
-    <script type="text/javascript" src="js/chromecast.js"></script>
   </body>
 </html>
